requirements.txt
fastapi
uvicorn
python-multipart
crewai[tools]
langchain
langchain-community
pypdf
scikit-learn
requests
langchain-groq
langchain-google-genai
python-dotenv
beautifulsoup4
faiss-cpu
PyPDF2
litellm
